BeliefPropagation

Documentation for BeliefPropagation.

BeliefPropagation.FactorGraphs.FactorGraph

BeliefPropagation.FactorGraphs.FactorGraph

BeliefPropagation.FactorGraphs.FactorGraphVertex

BeliefPropagation.FactorGraphs.factor

BeliefPropagation.FactorGraphs.factors

BeliefPropagation.FactorGraphs.nfactors

BeliefPropagation.FactorGraphs.nvariables

BeliefPropagation.FactorGraphs.pairwise_interaction_graph

BeliefPropagation.FactorGraphs.variable

BeliefPropagation.FactorGraphs.variables

Graphs.edges

Graphs.neighbors

IndexedGraphs.inedges

IndexedGraphs.outedges

BeliefPropagation.FactorGraphs.FactorGraph — Type

FactorGraph{T}

A type representing a factor graph.

BeliefPropagation.FactorGraphs.FactorGraph — Method

FactorGraph(A::AbstractMatrix)

Construct a FactorGraph from adjacency matrix A with the convention that rows are factors, columns are variables.

BeliefPropagation.FactorGraphs.FactorGraphVertex — Type

FactorGraphVertex

A type to represent a vertex in a bipartite graph, to be passed as an argument to neighbors, inedges, outedges, see examples therein. It is recommended to use the variable and factor constructors.

BeliefPropagation.FactorGraphs.factor — Method

factor(a::Integer)

Wraps index a in a container such that other functions like neighbors, inedges etc. know that it indices a factor node.

BeliefPropagation.FactorGraphs.factors — Method

factors(g::FactorGraph)

Return a lazy iterator to the indices of factor vertices in g.

BeliefPropagation.FactorGraphs.nfactors — Method

nactors(g::FactorGraph)

Return the number of actors vertices in g.

BeliefPropagation.FactorGraphs.nvariables — Method

nvariables(g::FactorGraph)

Return the number of variables vertices in g.

BeliefPropagation.FactorGraphs.pairwise_interaction_graph — Method

pairwise_interaction_graph(g::IndexedGraph)

Construct a factor graph whose factors are the pair-wise interactions encoded in g.

BeliefPropagation.FactorGraphs.variable — Method

variable(i::Integer)

Wraps index i in a container such that other functions like neighbors, inedges etc. know that it indices a variable node.

BeliefPropagation.FactorGraphs.variables — Method

variables(g::FactorGraph)

Return a lazy iterator to the indices of variable vertices in g.

Graphs.edges — Method

edges(g::FactorGraph)

Return a lazy iterator to the edges of g, with the convention that the source is the factor and the destination is the variable

Graphs.neighbors — Method

IndexedGraphs.neighbors(g::FactorGraph, v::FactorGraphVertex)

Return a lazy iterators to the neighbors of vertex v.

Examples

julia> using BeliefPropagation.FactorGraphs

julia> g = FactorGraph([0 1 1 0;
                        1 0 0 0;
                        0 0 1 1])
FactorGraph{Int64} with 3 factors, 4 variables and 5 edges

julia> collect(neighbors(g, variable(3)))
2-element Vector{Int64}:
 1
 3

julia> collect(neighbors(g, factor(2)))
1-element Vector{Int64}:
 1

IndexedGraphs.inedges — Method

IndexedGraphs.inedges(g::FactorGraph, v::FactorGraphVertex)

Return a lazy iterators to the edges incident on vertex v, with v as the destination.

Examples

julia> using BeliefPropagation.FactorGraphs

julia> g = FactorGraph([0 1 1 0;
                        1 0 0 0;
                        0 0 1 1])
FactorGraph{Int64} with 3 factors, 4 variables and 5 edges

julia> collect(inedges(g, factor(2)))
1-element Vector{IndexedGraphs.IndexedEdge{Int64}}:
 Indexed Edge 1 => 2 with index 1

julia> collect(inedges(g, variable(3)))
2-element Vector{IndexedGraphs.IndexedEdge{Int64}}:
 Indexed Edge 1 => 3 with index 3
 Indexed Edge 3 => 3 with index 4

IndexedGraphs.outedges — Method

IndexedGraphs.outedges(g::FactorGraph, v::FactorGraphVertex)

Return a lazy iterators to the edges incident on vertex v, with v as the source.

Examples

julia> using BeliefPropagation.FactorGraphs

julia> g = FactorGraph([0 1 1 0;
                        1 0 0 0;
                        0 0 1 1])
FactorGraph{Int64} with 3 factors, 4 variables and 5 edges

julia> collect(outedges(g, factor(2)))
1-element Vector{IndexedGraphs.IndexedEdge{Int64}}:
 Indexed Edge 2 => 1 with index 1

julia> collect(outedges(g, variable(3)))
2-element Vector{IndexedGraphs.IndexedEdge{Int64}}:
 Indexed Edge 3 => 1 with index 3
 Indexed Edge 3 => 3 with index 4