API reference

Docstrings

BeliefPropagation.BP — Type

BP{F<:BPFactor, FV<:BPFactor, M, MB, G<:FactorGraph}

A type representing the state of the Belief Propagation algorithm.

Fields

g: a [FactorGraph]

ψ: a vector of BPFactor representing the factors {ψₐ(xₐ)}ₐ

ϕ: a vector of BPFactor representing the single-variable factors {ϕᵢ(xᵢ)}ᵢ

u: messages from factor to variable

h: messages from variable to factor

b: beliefs

BeliefPropagation.BP — Method

BP(g::FactorGraph, ψ::AbstractVector{<:BPFactor}, states; ϕ)

Constructor for the BP type.

Arguments

g: a [FactorGraph]

ψ: a vector of BPFactor representing the factors {ψₐ(xₐ)}ₐ

states: an iterable of integers of length equal to the number of variable nodes specifyig the number of values each variable can take

ϕ: (optional) a vector of BPFactor representing the single-variable factors {ϕᵢ(xᵢ)}ᵢ

BeliefPropagation.BPFactor — Type

BPFactor

An abstract type representing a factor.

BeliefPropagation.TabulatedBPFactor — Type

TabulatedBPFactor

A type of BPFactor constructed by specifying the output to any input in a tabular fashion via an array values.

BeliefPropagation.UniformFactor — Type

UniformFactor

A type of BPFactor which returns the same value for any input: it behaves as if it wasn't even there. It is used as the default for single-variable factors

BeliefPropagation.avg_energy — Method

avg_energy([f], bp::BP)

Return the average energy

\[\langle E \rangle =\sum_a\sum_{\underline{x}_a}b_a(\underline{x}_a) \left[-\log\psi_a(\underline{x}_a)\right] + \sum_i\sum_{x_i}b_i(x_i) \left[-\log\phi_i(x_i)\right]\]

BeliefPropagation.beliefs — Method

beliefs([f], bp::BP)

Return single-variable beliefs $\{b_i(x_i)\}_i$.

BeliefPropagation.bethe_free_energy — Method

bethe_free_energy([f], bp::BP)

Return the bethe free energy

\[F=\sum_a\sum_{\underline{x}_a}b_a(\underline{x}_a) \left[-\log\frac{b_a(\underline{x}_a)}{\psi_a(\underline{x}_a)}\right] + \sum_i\sum_{x_i}b_i(x_i) \left[-\log\frac{b_i(x_i)^{1-\lvert\partial i\rvert}}{\phi_i(x_i)}\right]\]

BeliefPropagation.energy — Method

energy(bp::BP, x)

Return the energy

\[E(\underline{x})=\sum_a \left[-\log\psi_a(\underline{x}_a)\right] + \sum_i \left[-\log\phi_i(x_i)\right]\]

of configuration x.

BeliefPropagation.evaluate — Method

evaluate(bp::BP, x)

Return the unnormalized probability $\prod_a\psi_a(\underline{x}_a)\prod_i\phi_i(x_i)$ of configuration x.

BeliefPropagation.factor_beliefs — Method

factor_beliefs([f], bp::BP)

Return factor beliefs $\{b_a(\underline{x}_a)\}_a$.

BeliefPropagation.iterate! — Method

iterate!(bp::BP; kwargs...)

Run BP.

Optional arguments

update_variable!: the function that computes and updates variable-to-factor messages

update_factor!: the function that computes and updates factor-to-variable messages

maxiter: maximum number of iterations

tol: convergence check parameter

damp: damping parameter

rein: reinforcement parameter

callback

check_convergence: a function that checks if convergence has been reached

extra arguments to be passed to custom update_variable! and update_factor!

BeliefPropagation.iterate_ms! — Method

iterate_ms!(bp::BP; kwargs...)

Runs the max-sum algorithm (BP at zero temperature).

BeliefPropagation.nstates — Method

nstates(bp::BP, i::Integer)

Return the number of values taken by variable i.

BeliefPropagation.reset! — Method

reset!(bp::BP)

Reset all messages and beliefs to zero

BeliefPropagation.Models.fast_ising_bp — Function

fast_ising_bp(g::AbstractFactorGraph, ψ::Vector{<:IsingCoupling}, [ϕ])

Return a BP instance with Ising factors and messages and beliefs in log-ratio format:

\[\begin{align*} &m_{a\to i}(\sigma_i) \propto e^{u_{a\to i}\sigma_i}\\ &u_{a\to i} = \frac12\log\frac{m_{a\to i}(+1)}{m_{a\to i}(+1)} \end{align*}\]

BeliefPropagation.TabulatedBPFactor — Method

TabulatedBPFactor(f::BPFactor, states)

Construct a TabulatedBPFactor out of any BPFactor. Used mostly for tests.

BeliefPropagation.Test.exact_avg_energy — Method

exact_avg_energy(bp::BP; p_exact = exact_prob(bp))

Exhaustively compute the average energy (minus the log of the unnormalized probability weight).

BeliefPropagation.Test.exact_factor_marginals — Method

exact_factor_marginals(bp::BP; p_exact = exact_prob(bp))

Exhaustively compute marginal distributions for each factor.

BeliefPropagation.Test.exact_marginals — Method

exact_marginals(bp::BP; p_exact = exact_prob(bp))

Exhaustively compute marginal distributions for each variable.

BeliefPropagation.Test.exact_minimum_energy — Method

exact_minimum_energy(bp::BP)

Exhaustively compute the minimum energy (minus the log of the unnormalized probability weight).

BeliefPropagation.Test.exact_normalization — Method

exact_normalization(bp::BP)

Exhaustively compute the normalization constant.

BeliefPropagation.Test.exact_prob — Method

exact_prob(bp::BP; Z = exact_normalization(bp))

Exhaustively compute the probability of each possible configuration of the variables.

BeliefPropagation.Test.make_generic — Method

make_generic(bp::BP)

Return the corresponding BP with plain TabulatedBPFactor as factors. Used to test specific implementations.

BeliefPropagation.Test.rand_bp — Method

rand_bp([rng], g::FactorGraph, states)

Return a BP with random factors.

states is an iterable containing the number of values that can be taken by each variable.

BeliefPropagation.Test.rand_factor — Method

rand_factor([rng,], states)

Return a random BPFactor whose domain is specified by the iterable states.

BeliefPropagation.Test.test_observables_bp — Method

test_observables_bp(bp::BP; kwargs...)

Test beliefs_bp, factor_beliefs_bp and bethe_free_energy_bp against the same quantities computed exactly by exhaustive enumeration.

BeliefPropagation.Test.test_observables_bp_generic — Method

test_observables_bp_generic(bp::BP, bp_generic::BP; kwargs...)

Test beliefs_bp, factor_beliefs_bp and bethe_free_energy_bp against the same quantities on a generic version. See also make_generic

BeliefPropagation.Test.test_za — Method

test_za(ψₐ::BPFactor, states)

Test a specific implementation of compute_za against the naive one.

Index

BeliefPropagation.BP

BeliefPropagation.BP

BeliefPropagation.BPFactor

BeliefPropagation.TabulatedBPFactor

BeliefPropagation.TabulatedBPFactor

BeliefPropagation.UniformFactor

BeliefPropagation.Models.fast_ising_bp

BeliefPropagation.Test.exact_avg_energy

BeliefPropagation.Test.exact_factor_marginals

BeliefPropagation.Test.exact_marginals

BeliefPropagation.Test.exact_minimum_energy

BeliefPropagation.Test.exact_normalization

BeliefPropagation.Test.exact_prob

BeliefPropagation.Test.make_generic

BeliefPropagation.Test.rand_bp

BeliefPropagation.Test.rand_factor

BeliefPropagation.Test.test_observables_bp

BeliefPropagation.Test.test_observables_bp_generic

BeliefPropagation.Test.test_za

BeliefPropagation.avg_energy

BeliefPropagation.beliefs

BeliefPropagation.bethe_free_energy

BeliefPropagation.energy

BeliefPropagation.evaluate

BeliefPropagation.factor_beliefs

BeliefPropagation.iterate!

BeliefPropagation.iterate_ms!

BeliefPropagation.nstates

BeliefPropagation.reset!