BeliefPropagation

Documentation for BeliefPropagation.

BeliefPropagation.BP

BeliefPropagation.BP

BeliefPropagation.BPFactor

BeliefPropagation.FactorGraphs.FactorGraph

BeliefPropagation.FactorGraphs.FactorGraph

BeliefPropagation.FactorGraphs.FactorGraphVertex

BeliefPropagation.FactorGraphs.RegularFactorGraph

BeliefPropagation.TabulatedBPFactor

BeliefPropagation.TabulatedBPFactor

BeliefPropagation.UniformFactor

BeliefPropagation.FactorGraphs.edge_indices

BeliefPropagation.FactorGraphs.factor

BeliefPropagation.FactorGraphs.factors

BeliefPropagation.FactorGraphs.nfactors

BeliefPropagation.FactorGraphs.nvariables

BeliefPropagation.FactorGraphs.pairwise_interaction_graph

BeliefPropagation.FactorGraphs.rand_factor_graph

BeliefPropagation.FactorGraphs.rand_factor_graph

BeliefPropagation.FactorGraphs.rand_regular_factor_graph

BeliefPropagation.FactorGraphs.rand_tree_factor_graph

BeliefPropagation.FactorGraphs.variable

BeliefPropagation.FactorGraphs.variables

BeliefPropagation.avg_energy

BeliefPropagation.beliefs

BeliefPropagation.bethe_free_energy

BeliefPropagation.energy

BeliefPropagation.evaluate

BeliefPropagation.factor_beliefs

BeliefPropagation.iterate!

BeliefPropagation.iterate_ms!

BeliefPropagation.nstates

BeliefPropagation.rand_factor

BeliefPropagation.reset!

Graphs.edges

Graphs.neighbors

BeliefPropagation.BP — Type

BP{F<:BPFactor, FV<:BPFactor, M, MB, G<:FactorGraph}

A type representing the state of the Belief Propagation algorithm.

Fields

g: a FactorGraph

ψ: a vector of BPFactor representing the factors {ψₐ(xₐ)}ₐ

ϕ: a vector of BPFactor representing the single-variable factors {ϕᵢ(xᵢ)}ᵢ

u: messages from factor to variable

h: messages from variable to factor

b: beliefs

BeliefPropagation.BP — Method

BP(g::FactorGraph, ψ::AbstractVector{<:BPFactor}, qs; ϕ)

Constructor for the BP type.

Arguments

g: a FactorGraph

ψ: a vector of BPFactor representing the factors {ψₐ(xₐ)}ₐ

states: an iterable of integers of length equal to the number of variable nodes specifyig the number of values each variable can take

ϕ: (optional) a vector of BPFactor representing the single-variable factors {ϕᵢ(xᵢ)}ᵢ

BeliefPropagation.BPFactor — Type

BPFactor

An abstract type representing a factor.

BeliefPropagation.TabulatedBPFactor — Type

TabulatedBPFactor

A type of BPFactor constructed by specifying the output to any input in a tabular fashion via an array values.

BeliefPropagation.TabulatedBPFactor — Method

TabulatedBPFactor(f::BPFactor, states)

Construct a TabulatedBPFactor out of any BPFactor. Used mostly for tests.

BeliefPropagation.UniformFactor — Type

UniformFactor

A type of BPFactor which returns the same value for any input: it behaves as if it wasn't even there. It is used as the default for single-variable factors

BeliefPropagation.avg_energy — Method

avg_energy([f], bp::BP)

Return the average energy ∑ₐ∑ₓₐbₐ(xₐ)[-logψₐ(xₐ)] + ∑ᵢ∑ₓᵢbᵢ(xᵢ)[-logϕᵢ(xᵢ)]

BeliefPropagation.beliefs — Method

beliefs([f], bp::BP)

Return single-variable beliefs {bᵢ(xᵢ)}ᵢ.

BeliefPropagation.bethe_free_energy — Method

bethe_free_energy([f], bp::BP)

Return the bethe free energy ∑ₐ∑ₓₐbₐ(xₐ)log[bₐ(xₐ)/ψₐ(xₐ)] + ∑ᵢ∑ₓᵢbᵢ(xᵢ)log[bᵢ(xᵢ)^(1-|∂i|)/ϕᵢ(xᵢ)]

BeliefPropagation.energy — Method

energy(bp::BP, x)

Return the energy ∑ₐ[-logψₐ(xₐ)] + ∑ᵢ[-logϕᵢ(xᵢ)] of configuration x.

BeliefPropagation.evaluate — Method

evaluate(bp::BP, x)

Return the unnormalized probability ∏ₐψₐ(xₐ)∏ᵢϕᵢ(xᵢ) of configuration x.

BeliefPropagation.factor_beliefs — Method

factor_beliefs([f], bp::BP)

Return factor beliefs {bₐ(xₐ)}ₐ.

BeliefPropagation.iterate! — Method

iterate!(bp::BP; kwargs...)

Run BP.

Optional arguments

update_variable!: the function that computes and updates variable-to-factor messages

update_factor!: the function that computes and updates factor-to-variable messages

maxiter: maximum number of iterations

tol: convergence check parameter

damp: damping parameter

rein: reinforcement parameter

f: a vector to store on-the-fly computations of the bethe free energy

callback

check_convergence: a function that checks if convergence has been reached

extra arguments to be passed to custom update_variable! and update_factor!

BeliefPropagation.iterate_ms! — Method

iterate_ms!(bp::BP; kwargs...)

Runs the max-sum algorithm (BP at zero temperature).

BeliefPropagation.nstates — Method

nstates(bp::BP, i::Integer)

Return the number of values taken by variable i.

BeliefPropagation.rand_factor — Method

rand_factor([rng,], states)

Return a random BPFactor whose domain is specified by the iterable states.

Examples

Create a random factor connected to three variables x₁ ∈ {1,2}, x₂ ∈ {1,2,3,4}, x₃ ∈ {1,2,3}.

julia> using BeliefPropagation, BeliefPropagation.FactorGraphs

julia> import Random: MersenneTwister

julia> states = (2, 4, 3);

julia> f = rand_factor(MersenneTwister(0), states);

julia> f([1, 4, 2])
0.16703619444214968

BeliefPropagation.reset! — Method

reset!(bp::BP)

Reset all messages and beliefs to zero

BeliefPropagation.FactorGraphs.FactorGraph — Type

FactorGraph{T}

A type representing a factor graph.

BeliefPropagation.FactorGraphs.FactorGraph — Method

FactorGraph(A::AbstractMatrix)

Construct a FactorGraph from adjacency matrix A with the convention that rows are factors, columns are variables.

BeliefPropagation.FactorGraphs.FactorGraphVertex — Type

FactorGraphVertex

A type to represent a vertex in a bipartite graph, to be passed as an argument to neighbors, see examples therein. It is recommended to use the variable and factor constructors.

BeliefPropagation.FactorGraphs.RegularFactorGraph — Type

RegularFactorGraph <: AbstractFactorGraph

A type to represent an infinite regular factor graph with fixed factor and variable degree

BeliefPropagation.FactorGraphs.edge_indices — Method

IndexedGraphs.edge_indices(g::FactorGraph, v::FactorGraphVertex)

Return a lazy iterators to the indices of the edges incident on vertex v, with v.

The output of edge_indices does not allocate and it can be used to index external arrays of properties directly

Examples

julia> using BeliefPropagation.FactorGraphs, Random

julia> g = FactorGraph([0 1 1 0;
                        1 0 0 0;
                        0 0 1 1])
FactorGraph{Int64} with 3 factors, 4 variables and 5 edges

julia> edgeprops = randn(MersenneTwister(0), ne(g));

julia> indices = edge_indices(g, variable(3));

julia> edgeprops[indices]
2-element Vector{Float64}:
 -0.3530074003005963
 -0.13485387193052173

BeliefPropagation.FactorGraphs.factor — Method

factor(a::Integer)

Wraps index a in a container such that other functions like neighbors know that it indices a factor node.

BeliefPropagation.FactorGraphs.factors — Method

factors(g::FactorGraph)

Return a lazy iterator to the indices of factor vertices in g.

BeliefPropagation.FactorGraphs.nfactors — Method

nactors(g::FactorGraph)

Return the number of actors vertices in g.

BeliefPropagation.FactorGraphs.nvariables — Method

nvariables(g::FactorGraph)

Return the number of variables vertices in g.

BeliefPropagation.FactorGraphs.pairwise_interaction_graph — Method

pairwise_interaction_graph(g::IndexedGraph)

Construct a factor graph whose factors are the pair-wise interactions encoded in g.

BeliefPropagation.FactorGraphs.rand_factor_graph — Method

rand_factor_graph([rng=default_rng()], nvar, nfact, ned)

Create a factor graph with nvar variables, nfact factors and ned edges taken uniformly at random.

BeliefPropagation.FactorGraphs.rand_factor_graph — Method

rand_factor_graph([rng=default_rng()], nvar, nfact, p)

Create a factor graph with nvar variables, nfact factors and edges taken independently with probability p.

BeliefPropagation.FactorGraphs.rand_regular_factor_graph — Method

rand_regular_factor_graph([rng=default_rng()], nvar, nfact, k)

Create a factor graph with nvar variables and nfact factors, where all factors have degree k.

BeliefPropagation.FactorGraphs.rand_tree_factor_graph — Method

rand_tree_factor_graph([rng=default_rng()], n)

Create a tree factor graph with n vertices in total. The proportion of variables/factors is casual.

BeliefPropagation.FactorGraphs.variable — Method

variable(i::Integer)

Wraps index i in a container such that other functions like neighbors know that it indices a variable node.

BeliefPropagation.FactorGraphs.variables — Method

variables(g::FactorGraph)

Return a lazy iterator to the indices of variable vertices in g.

Graphs.edges — Method

edges(g::FactorGraph)

Return a lazy iterator to the edges of g, with the convention that the source is the factor and the destination is the variable

julia> using BeliefPropagation.FactorGraphs

julia> g = FactorGraph([0 1 1 0;
                        1 0 0 0;
                        0 0 1 1])
FactorGraph{Int64} with 3 factors, 4 variables and 5 edges

julia> collect(edges(g))
5-element Vector{IndexedGraphs.IndexedEdge{Int64}}:
 Indexed Edge 2 => 1 with index 1
 Indexed Edge 1 => 2 with index 2
 Indexed Edge 1 => 3 with index 3
 Indexed Edge 3 => 3 with index 4
 Indexed Edge 3 => 4 with index 5

Graphs.neighbors — Method

IndexedGraphs.neighbors(g::FactorGraph, v::FactorGraphVertex)

Return a lazy iterators to the neighbors of vertex v.

Examples

julia> using BeliefPropagation.FactorGraphs

julia> g = FactorGraph([0 1 1 0;
                        1 0 0 0;
                        0 0 1 1])
FactorGraph{Int64} with 3 factors, 4 variables and 5 edges

julia> collect(neighbors(g, variable(3)))
2-element Vector{Int64}:
 1
 3

julia> collect(neighbors(g, factor(2)))
1-element Vector{Int64}:
 1